requirements.txt
chainlit>=1.0.0
uvicorn
gunicorn
python-dotenv
azure-ai-projects
azure-identity